Moved  test to signing.rs for dependencies issues
